Custom LeftNavMenu component + its hook up to Redux to show dynamic list of classes
QcDrawerItem for list items in the drawer menu. It sitll needs to be generalized to accept icons as well as images for avatars
Checked in few icons for classes. These are placeholders until we get official ones.
Next: need to provide a way in add class screen for teachers to change / customize their class image.