Change tracking screen names logic
I started seeing in the telemetry some weird names of loaded screens, and after investigation it turns out that we can't rely to this.constructor.name to capture the screen name.

We used to rely on contructor.name to automatically get the name of the loaded screen for analytics. Unfortunately according to ES6 specs this only works on non compiled JS.

I switched to explicity setting a varialbe with the name of each screen.