Fixes and enhancements to the latest changes
The new assignmentEvaluation Screen always showed current assignment name (while it showed th grade and notes of a past assignment). Fixed it to show the proper assignment name

Added assignment completion date to the new assignment evaluation screen.

Changed the placeholder string "Write a note" on the past evaluation screen, since it is read only for that case. The new string is: "There are no additional notes provided"

Invoke the new assignment evaluation card when user presses on any area within the past assignment card. (It was only invoked when user presses on the notes section, which was not very discoverable)